# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/menu/items": {
            "post": {
                "description": "Create a new menu item with name, price, description and optional image/logo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Create a new menu item",
                "parameters": [
                    {
                        "description": "Menu item data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menuitemapp.CreateMenuItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create menu item success",
                        "schema": {
                            "$ref": "#/definitions/app.CreateMenuItemSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/menu/items/search": {
            "post": {
                "description": "Get list of menu items by restaurant ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Get menu items by restaurant",
                "parameters": [
                    {
                        "description": "Search menu items request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menuitemapp.GetMenuItemsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get menu items by restaurant successfully",
                        "schema": {
                            "$ref": "#/definitions/app.GetMenuItemsByRestaurantSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/menu/items/{id}": {
            "get": {
                "description": "Get a menu item detail by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Get menu item by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Menu item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get menu item successfully",
                        "schema": {
                            "$ref": "#/definitions/app.GetMenuItemSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            },
            "put": {
                "description": "Update menu item by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Update menu item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Menu item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update menu item data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menuitemapp.UpdateMenuItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update menu item successfully",
                        "schema": {
                            "$ref": "#/definitions/app.UpdateMenuItemSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete menu item by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Delete menu item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Menu item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete menu item successfully",
                        "schema": {
                            "$ref": "#/definitions/app.DeleteMenuItemSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/menu/items/{id}/status": {
            "patch": {
                "description": "Enable or disable menu item by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Update menu item status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update menu item status request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menuitemapp.UpdateMenuItemStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update menu item status successfully",
                        "schema": {
                            "$ref": "#/definitions/app.UpdateMenuItemStatusSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.CreateMenuItemSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.DeleteMenuItemSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.GetMenuItemSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/menuitemapp.GetMenuItemResponse"
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
        "app.GetMenuItemsByRestaurantResponseDoc": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/menuitemapp.GetMenuItemResponse"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total_items": {
                    "type": "integer",
                    "example": 42
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "app.GetMenuItemsByRestaurantSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/app.GetMenuItemsByRestaurantResponseDoc"
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.UpdateMenuItemStatusSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
        "app.UpdateMenuItemSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "domain.MenuItemType": {
            "type": "string",
            "enum": [
                "dish",
                "beverage",
                "extra",
                "combo"
            ],
            "x-enum-varnames": [
                "ItemDish",
                "ItemDrink",
                "ItemExtra",
                "ItemCombo"
            ]
        },
        "menuitemapp.CreateMenuItemRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "topic_id": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/domain.MenuItemType"
                }
            }
        },
        "menuitemapp.GetMenuItemResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/domain.MenuItemType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "menuitemapp.GetMenuItemsRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "filter": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "menuitemapp.UpdateMenuItemRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/domain.MenuItemType"
                }
            }
        },
        "menuitemapp.UpdateMenuItemStatusRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                }
            }
        },
...
        "response.ErrorDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
    }
}